Refactor event normalization into NormalizerPipeline
Extracted basic text normalization and location enrichment logic from `shared-core.js` into dedicated `BasicDataNormalizer` and `LocationNormalizer` classes within a new `normalizers.js` file.

Updated the orchestrator (`bear-event-scraper-unified.js`) to load, instantiate, and invoke the pipeline immediately after parsing and after deduplication/merging, ensuring `event.notes` are properly regenerated.
