+++ img/favicons/favicon-seattleeagle.com-256px.ico.meta
@@ -1,9 +1,9 @@
 {
   "originalUrl": "https://www.google.com/s2/favicons?domain=seattleeagle.com&sz=256",
-  "failedAt": "2026-06-16T13:04:02.550Z",
+  "failedAt": "2026-06-23T18:23:31.712Z",
   "error": "HTTP 404: Not Found",
   "type": "favicon",
   "filename": "favicon-seattleeagle.com-256px.ico",
   "size": "256",
-  "failureCount": 1
+  "failureCount": 2
 }
