+++ img/favicons/favicon-seattleeagle.com-64px.ico.meta
@@ -1,9 +1,9 @@
 {
   "originalUrl": "https://www.google.com/s2/favicons?domain=seattleeagle.com&sz=64",
-  "failedAt": "2026-06-16T13:04:02.469Z",
+  "failedAt": "2026-06-23T18:23:31.672Z",
   "error": "HTTP 404: Not Found",
   "type": "favicon",
   "filename": "favicon-seattleeagle.com-64px.ico",
   "size": "64",
-  "failureCount": 1
+  "failureCount": 2
 }
