+++ package.json
@@ -30,6 +30,6 @@
     "express": "^5.1.0",
     "playwright": "^1.61.0",
     "puppeteer": "^22.15.0",
-    "sharp": "^0.33.0"
+    "sharp": "^0.33.5"
   }
 }
